Add files via upload
This is a script meant to tell the time. To run simply type tellTime in the command line for powershell and it will output the time. I learned more about the get-date cmdlet from https://blog.itpro.tv/get-date-powershell-cmdlet/ and incorporated this into my project. It is simple but I want to expand on this to do something complex with telling the time/date.